// src/navigation/types.ts
export type HomeStackParamList = {
  Home: undefined;
  WealthJourney: undefined;
};

export type SaveStackParamList = {
  Save: undefined;
  SipDetails: undefined;
  BankLink: undefined;
};

export type BottomTabParamList = {
  HomeTab: undefined;
  SaveTab: undefined;
  BorrowTab: undefined;
  RewardsTab: undefined;
};